Effectful.Environment.Guard.Static

Class

class Monad m => MonadEnvGuard m where

Since: 0.1

Methods

guardPredicate :: String -> (String -> Bool) -> m a -> m (Maybe a)

Since: 0.1

Effect

data EnvGuardStatic :: Effect

Static effect for EnvGuard.

Since: 0.1

Handler

runEnvGuardStaticIO :: IOE :> es => Eff (EnvGuardStatic : es) a -> Eff es a

Runs an EnvGuardStatic effect.

Since: 0.1

High level combinators

data ExpectEnv

The expectation for an environment variable lookup.

Since: env-guard-0.1.1

Constructors

ExpectEnvSet
Expect that the environment variable is set (i.e. contents can be anything).

Since: env-guard-0.1.1

ExpectEnvEquals String
Expect that the environment variable is set and the contents equals the string. This is case-insensitive.

Since: env-guard-0.1.1

ExpectEnvPredicate (String -> Bool)
Expect that the environment variable is set and its contents satisfies the predicate.

Since: env-guard-0.1.1

withGuard :: MonadEnvGuard m => String -> ExpectEnv -> m a -> m (Maybe a)

Since: 0.1

withGuard_ :: MonadEnvGuard m => String -> ExpectEnv -> m a -> m ()

Since: 0.1

guardOrElse

Arguments

:: MonadEnvGuard m

=> String
The environment variable.

-> ExpectEnv
The expectation.

-> m a
The action to run if the expectation succeeds.

-> m e
The action to run if the expectation fails.

-> m (Either e a)
The result.

Since: 0.1

guardOrElse'

Arguments

:: MonadEnvGuard m

=> String
The environment variable.

-> ExpectEnv
The expectation.

-> m a
The action to run if the expectation succeeds.

-> m a
The action to run if the expectation fails.

-> m a
The result.

Since: 0.1

Low level functions

Checking environment variable is set

guardSet :: MonadEnvGuard m => String -> m a -> m (Maybe a)

Since: 0.1

guardSet_ :: MonadEnvGuard m => String -> m a -> m ()

Since: 0.1

Checking environment variable match

guardEquals :: MonadEnvGuard m => String -> String -> m a -> m (Maybe a)

Since: 0.1

guardEquals_ :: MonadEnvGuard m => String -> String -> m a -> m ()

Since: 0.1

Checking environment variable predicate

guardPredicate :: MonadEnvGuard m => String -> (String -> Bool) -> m a -> m (Maybe a)

Since: 0.1

guardPredicate_ :: MonadEnvGuard m => String -> (String -> Bool) -> m a -> m ()

Since: 0.1